app/back/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0

# Database
psycopg2-binary==2.9.9

# Development
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Documentation
python-multipart==0.0.6
